Changed the Tools>Options dialog to always reload from the saved configuration file on [ OK ]. Plus minor text tweak for the auto touch keyboard option.

+++ Epi Info 7/Epi.Windows.Enter/Forms/Canvas.control.events.cs
@@ -526,7 +526,10 @@ namespace Epi.Windows.Enter
         {
             try 
             { 
-                keyboardProcess.Kill(); 
+                if(keyboardProcess.HasExited == false)
+                { 
+                    keyboardProcess.Kill();
+                }
             }
             catch { }
